delete the redundant file and add train model
